Fix: 예약 관리 리스트에서 신한그룹사 여부가 항상 false로 반환되는 문제 해결
- buildIsShinhanByUserId()의 userId 키 타입을 Long으로 통일
- Math.toIntExact() 변환 제거로 키 불일치 문제 수정
- Controller에서 Long 타입 키로 정상 조회 가능하도록 반영

+++ src/main/java/Team_Mute/back_end/domain/reservation_admin/service/ReservationAdminService.java
@@ -280,7 +280,7 @@ public class ReservationAdminService {
 			.collect(Collectors.toMap(UserCompany::getCompanyId, UserCompany::getCompanyName));
 
 		// userId -> isShinhan 맵
-		Map<Integer, Boolean> isShinhanByUserId =
+		Map<Long, Boolean> isShinhanByUserId =
 			ShinhanGroupUtils.buildIsShinhanByUserId(users, companyNameById);
 
 		// DTO 변환
